# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/appointments": {
      "post": {
        "operationId": "AppointmentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The appointment has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new appointment (Admin will do it)",
        "tags": [
          "appointments"
        ]
      },
      "get": {
        "operationId": "AppointmentsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (starts from 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns all appointments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppointmentResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all appointments",
        "tags": [
          "appointments"
        ]
      }
    },
    "/appointments/{id}": {
      "get": {
        "operationId": "AppointmentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Appointment ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the appointment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Appointment not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get an appointment by ID",
        "tags": [
          "appointments"
        ]
      },
      "put": {
        "operationId": "AppointmentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Appointment ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The appointment has been successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request."
          },
          "404": {
            "description": "Appointment not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an appointment",
        "tags": [
          "appointments"
        ]
      },
      "delete": {
        "operationId": "AppointmentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Appointment ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The appointment has been successfully deleted."
          },
          "404": {
            "description": "Appointment not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an appointment",
        "tags": [
          "appointments"
        ]
      }
    },
    "/appointments/patient/{patientId}": {
      "get": {
        "operationId": "AppointmentsController_findByPatient",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "description": "Patient user ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns appointments for the patient",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppointmentResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get appointments by patient ID",
        "tags": [
          "appointments"
        ]
      }
    },
    "/appointments/therapist/{therapistId}": {
      "get": {
        "operationId": "AppointmentsController_findByTherapist",
        "parameters": [
          {
            "name": "therapistId",
            "required": true,
            "in": "path",
            "description": "Therapist user ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns appointments for the therapist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppointmentResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get appointments by therapist ID",
        "tags": [
          "appointments"
        ]
      }
    },
    "/appointments/request-appointment": {
      "post": {
        "operationId": "AppointmentsController_requestAppointment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The appointment has been successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Request an appointment (Intended to be used by patient)",
        "tags": [
          "appointments"
        ]
      }
    },
    "/appointments/confirm-appointment/{id}": {
      "put": {
        "operationId": "AppointmentsController_confirmAppointment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Appointment ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The appointment has been successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirm an appointment (Intended to be used by admin)",
        "tags": [
          "appointments"
        ]
      }
    },
    "/appointments/notes": {
      "post": {
        "operationId": "TherapistNotesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTherapistNoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TherapistNote"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Create a therapist note",
        "tags": [
          "Therapist Notes"
        ]
      }
    },
    "/appointments/notes/{appointmentId}": {
      "get": {
        "operationId": "TherapistNotesController_findByAppointment",
        "parameters": [
          {
            "name": "appointmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TherapistNote"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Get therapist notes for an appointment",
        "tags": [
          "Therapist Notes"
        ]
      }
    },
    "/appointments/notes/{id}": {
      "put": {
        "operationId": "TherapistNotesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTherapistNoteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TherapistNote"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Update a therapist note",
        "tags": [
          "Therapist Notes"
        ]
      },
      "delete": {
        "operationId": "TherapistNotesController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a therapist note",
        "tags": [
          "Therapist Notes"
        ]
      }
    },
...
  },
  "info": {
    "title": "Therapist Booking System API",
    "description": "API documentation for the therapist booking system",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateAppointmentDto": {
        "type": "object",
        "properties": {
          "therapyTypeId": {
            "type": "number",
            "description": "ID of the therapy type",
            "example": 1
          },
          "duration": {
            "type": "number",
            "description": "Duration of the appointment in minutes",
            "example": 60
          },
          "scheduledAt": {
            "type": "string",
            "description": "Scheduled date and time of the appointment",
            "example": "2025-04-01T14:00:00Z"
          },
          "isCompleted": {
            "type": "boolean",
            "description": "Whether the appointment is completed",
            "example": false,
            "default": false
          },
          "meetingUrl": {
            "type": "string",
            "description": "URL for the meeting/video call",
            "example": "https://zoom.us/j/123456789"
          },
          "patientIds": {
            "description": "Array of patient user IDs",
            "example": [
              1,
              2
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "therapistIds": {
            "description": "Array of therapist user IDs",
            "example": [
              3,
              4
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "therapyTypeId",
          "duration",
          "scheduledAt",
          "patientIds",
          "therapistIds"
        ]
      },
      "TherapyType": {
        "type": "object",
        "properties": {}
      },
      "PatientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com"
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "email"
        ]
      },
      "TherapistDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 3
          },
          "firstName": {
            "type": "string",
            "example": "Jane"
          },
          "lastName": {
            "type": "string",
            "example": "Smith"
          },
          "email": {
            "type": "string",
            "example": "jane.smith@example.com"
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "email"
        ]
      },
      "AppointmentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "duration": {
            "type": "number",
            "example": 60
          },
          "scheduledAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-04-01T14:00:00Z"
          },
          "isCompleted": {
            "type": "boolean",
            "example": false
          },
          "meetingUrl": {
            "type": "string",
            "example": "https://zoom.us/j/123456789"
          },
          "therapyType": {
            "$ref": "#/components/schemas/TherapyType"
          },
          "patients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PatientDto"
            }
          },
          "therapists": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TherapistDto"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-03-24T12:00:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-03-24T12:00:00Z"
          },
          "isConfirmed": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "id",
          "duration",
          "scheduledAt",
          "isCompleted",
          "therapyType",
          "patients",
          "therapists",
          "createdAt",
          "updatedAt",
          "isConfirmed"
        ]
      },
      "UpdateAppointmentDto": {
        "type": "object",
        "properties": {
          "therapyTypeId": {
            "type": "number",
            "description": "ID of the therapy type",
            "example": 1
          },
          "duration": {
            "type": "number",
            "description": "Duration of the appointment in minutes",
            "example": 60
          },
          "scheduledAt": {
            "type": "string",
            "description": "Scheduled date and time of the appointment",
            "example": "2025-04-01T14:00:00Z"
          },
          "isCompleted": {
            "type": "boolean",
            "description": "Whether the appointment is completed",
            "example": true,
            "default": false
          },
          "meetingUrl": {
            "type": "string",
            "description": "URL for the meeting/video call",
            "example": "https://zoom.us/j/123456789"
          },
          "patientIds": {
            "description": "Array of patient user IDs",
            "example": [
              1,
              2
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "therapistIds": {
            "description": "Array of therapist user IDs",
            "example": [
              3,
              4
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        }
      },
      "RequestAppointmentDto": {
        "type": "object",
        "properties": {
          "therapyTypeId": {
            "type": "number",
            "description": "ID of the therapy type",
            "example": 1
          },
          "duration": {
            "type": "number",
            "description": "Duration of the appointment in minutes",
            "example": 60
          },
          "scheduledAt": {
            "format": "date-time",
            "type": "string",
            "description": "Scheduled date and time of the appointment",
            "example": "2025-04-01T14:00:00Z"
          }
        },
        "required": [
          "therapyTypeId",
          "duration",
          "scheduledAt"
        ]
      },
      "ConfirmAppointmentDto": {
        "type": "object",
        "properties": {
          "therapistId": {
            "type": "number",
            "description": "ID of the therapist",
            "example": 42
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "description": "Appointment date and time with timezone",
            "example": "2023-12-25T14:30:00.000Z"
          }
        },
        "required": [
          "therapistId",
          "date"
        ]
      },
      "CreateTherapistNoteDto": {
        "type": "object",
        "properties": {
          "appointmentId": {
            "type": "number",
            "example": 1,
            "description": "Appointment ID"
          },
          "note": {
            "type": "string",
            "example": "Patient responded well to the session.",
            "description": "Therapist note"
          },
          "isVisibleToPatient": {
            "type": "boolean",
            "example": false,
            "description": "Is note visible to user"
          }
        },
        "required": [
          "appointmentId",
          "note",
          "isVisibleToPatient"
        ]
      },
      "TherapistNote": {
        "type": "object",
        "properties": {}
      },
      "UpdateTherapistNoteDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "example": "Updated note content",
            "description": "Updated therapist note"
          },
          "isVisibleToPatient": {
            "type": "boolean",
            "example": false,
            "description": "Is note visible to user"
          }
        },
        "required": [
          "note",
          "isVisibleToPatient"
        ]
      },
...
    }
  }
}